Feat: Authorization for admin

+++ src/main/java/com/example/SelfOrderingRestaurant/Dto/Response/DishResponseDTO/DishResponseDTO.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class DishResponseDTO {
     private String name;
-    private Long price;
+    private BigDecimal price;
     private Integer categoryId;
     private DishStatus status;
 }
